refactor: split dnsmasq_manager into modular components
Split the monolithic dnsmasq_manager.py into smaller, focused modules:
- base.py: Base class with common file writing functionality
- interface_handler.py: Virtual interface operations
- dhcp_config.py: DHCP configuration generation and ranges
- manager_mode.py: Manager mode specific logic
- metalbox_mode.py: Metalbox mode specific logic
- manager.py: Main orchestrator class

AI-assisted: Claude Code
