+++ files/netbox/dnsmasq/__init__.py
@@ -1,0 +1,19 @@
+# SPDX-License-Identifier: Apache-2.0
+
+"""DNS and DHCP configuration management modules."""
+
+from .base import DnsmasqBase
+from .dhcp_config import DHCPConfigGenerator
+from .interface_handler import InterfaceHandler
+from .manager_mode import ManagerModeHandler
+from .metalbox_mode import MetalboxModeHandler
+from .manager import DnsmasqManager
+
+__all__ = [
+    "DnsmasqBase",
+    "DHCPConfigGenerator",
+    "InterfaceHandler",
+    "ManagerModeHandler",
+    "MetalboxModeHandler",
+    "DnsmasqManager",
+]
